requirements.txt
numpy
pandas
pdfminer.six
unidecode